Adds ability to select tag manager
Note that limited user accounts are not permitted to call the respective
HTTP API method, even if the tag manager is already selected. Presumably
this is because limited user accounts are on a per tag manager basis.

In most cases this won't matter, because the respective tag manager is
selected to start with for a limited user account, so the HTTP API method
won't be called. I don't know whether a limited user account could actually
have multiple tag managers associated with them, in which case only one
would be selected.