Removes updateUntil() method
This wasn't currently being used, had a poorly designed call signature,
and it hasn't been clear what this would be used for that would be a
clear advantage. The few places where retries are taking place they are
easy (one attempt, not matter what), and so for those this method wasn't
useful (and hence not used).

Maybe there are cases later that merit bringing it back.