Tweaks the logic for computing timeout delay for update loop
Specifically, removes the default padding (this wasn't truly needed
strictly speaking), and introduces the notion of a delay with which
tag data updates are available through the cloud API. This turns out
to be significant (almost one minute), and hence not accounting for
it made for excessively many update() calls in short order because
the lastUpdated() status wasn't updating, which especially for
shorter update intervals seemed to cause erroneously short intervals
between recorded updates.

This value may well be different between different instances of the
cloud, but I don't have tag managers on different cloud instances, so
can't check that out.