Changes transform for mapped objects to enforce properties.
The behaviors that this changes are those in response of likely programming
errors for properties whose values are objects (dicts). Specifically, now
the property itself is read-only, preventing assignment of an object (that
could easily be malformed), and instead the object is returned with get/set
for each mapped property. Accidental creation of additional properties by
mistyping object property names will now result in a TypeError (if 'strict'
is enabled).