Increments version to v0.5.0
There is very little new functionality compared to v0.4.x, but the behavior
of a number of API functions is tightened up and made consistent as a result
of more comprehensive testing.

Also, testing is now divided into a suite that executes on CI but is
(mostly) read-only so as not to interfere with tags actually in use, and
a suite that runs through operations that directly alter tag state and/or
behavior. The latter is not run on CI, but is run locally.