Makes callback behavior more consistent across API calls.
I've been treating callbacks as a stepchild here in favor of Promises.
This will in general remain so, as Promises allow a much cleaner flow
of programming with a series of interdependent API calls, but this
is an attempt to make callback behavior at least a little more
consistent across the board, and more consistent with what the promises
return.

Callback functions for JSON API-calling methods now all get passed an
object with keys object (the object on which the operation was performed),
and value (the resulting value of the operation, if different from the
object).

Generally, it is a bad idea to use both callbacks and the returned promises
simultaneously. Sometimes it may work, but it is not currently and never
will be supported.