Adds callback to polling tag updater.
The use case here is allowing client apps to be informed about the full
set of data received from the API endpoint, regardless of which tag objects
are registered for receicing updates.

This may need to be changed to an event emitter interface in the future.
However, for now I wanted to keep the footprint of interface expectations
for tag updater classes low, and the fact that a callback argument is
ephemeral between start/stop cycles, whereas a registered event handler
persists (and thus needs code to prevent duplication), may be an advantage
rather than a problem.