Allows tag discovery in one go from platform object
The API endpoints being called for the all-in-one-go discovery are really
the same as for the one-tag-manager-at-a-time path. Therefore, while there
is little difference if there is only a single tag manager, for those with
multiple tag managers the all-in-one-go path should be better performing
because each API endpoint (one for finding tag managers, one for finding
tags) is only called once, and it's the code that stitches their results
together.

Also includes tests for the new methods.