Allows X-Set-Mac HTTP header instead of select API call
This header is undocumented in the [JSON API documentation], but is
what the web-application (see under "Example code") uses. If there is only
once tag manager accessible to the account, this should make no difference,
but for accounts with multiple tag managers switching between them from one
API call to the next should be faster, as it reduces two API calls to one.

This change will also support other custom headers etc in the future.

[JSON API documentation]: http://wirelesstag.net/media/mytaglist.com/apidoc.html