Adds recognition and handling for Outdoor series tags and sensors
This includes new and updated API methods for querying a tag's capabilities.
Also introduces a "secondary" temperature sensor for those tags can use _and_
make available the readings from more than one temperature sensor.

Currently _only_ the "Thermocouple" model in the Outdoor series does this:
the one treated primary is the external probe, and the secondary is an the
on-board chip. Note that although in theory the "GE Protimeter" model in the
Outdoor series _also_ measures temperature both from an (optional) external
DS18B20 probe and with an on-board chip, only one reading is available (the
probe if activated, and the internal on-board chip otherwise).