Deprecates tag.isHTU() from the API
The name of this method was originally borrowed from the Wireless Sensor Tags
Web SDK code, where the meaning was never made explicit. It was used both for
querying precision of temperature reading _and_ having a humidity sensor.
Until the arrival of the Outdoor tag models, these used to be synonymous (tags
with 8bit precision temperature did not have a humidity sensor). Now, however,
the Outdoor tag models all support high-precision temperature reading (even if
less than 13bit), yet only one of them can also measure humidity/moisture.

Here in the API we have been using (and documenting) isHTU() in the semantics
of having high-precision temperature. To fully disambiguate, we are now
deprecating isHTU() as an alias for the preferable tag.canHighPrecTemp().