Fixes erroneous provision of probeType for water sensors
This seems to have been introduced as part of b7fdebac3829f35f54df583de5ad4887c744b8d2.
It's not clear why this inconsistency between the sensor property definitions
here and the tests (which say water sensors aren't among the ones that
should have a value other than undefined for this property) wasn't caught
earlier. Between changing the tests and taking back this part of the earlier
change, I decided on the latter, which makes behavior consistent with the
tests _and_ with the documentation for the sensor class.